+++ tests/test_fetch.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-"""Test jumping within tabix files
+"""Test fetching within tabix files
 """
 
 import os
@@ -11,22 +11,22 @@ from vcfpy import reader
 __author__ = 'Manuel Holtgrewe <manuel.holtgrewe@bihealth.de>'
 
 
-def test_jump_no_records():
+def test_fetch_no_records():
     path = os.path.join(os.path.dirname(__file__), 'vcfs',
                         'multi_contig.vcf.gz')
     r = reader.Reader.from_path(path)
 
-    records = [vcf_rec for vcf_rec in r.jump_to('20', 1110698, 1230236)]
+    records = [vcf_rec for vcf_rec in r.fetch('20', 1110698, 1230236)]
 
     assert len(records) == 0
 
 
-def test_jump_one_record():
+def test_fetch_one_record():
     path = os.path.join(os.path.dirname(__file__), 'vcfs',
                         'multi_contig.vcf.gz')
     r = reader.Reader.from_path(path)
 
-    records = [vcf_rec for vcf_rec in r.jump_to('20', 1110695, 1230236)]
+    records = [vcf_rec for vcf_rec in r.fetch('20', 1110695, 1230236)]
 
     assert len(records) == 1
 
@@ -34,12 +34,12 @@ def test_jump_one_record():
     assert records[0].POS == 1110696
 
 
-def test_jump_two_records():
+def test_fetch_two_records():
     path = os.path.join(os.path.dirname(__file__), 'vcfs',
                         'multi_contig.vcf.gz')
     r = reader.Reader.from_path(path)
 
-    records = [vcf_rec for vcf_rec in r.jump_to('20', 1110697, 1234568)]
+    records = [vcf_rec for vcf_rec in r.fetch('20', 1110697, 1234568)]
 
     assert len(records) == 2