+++ tests/test_parser_record.py
@@ -24,13 +24,14 @@ MEDIUM_HEADER = """
 ##INFO=<ID=AA,Number=1,Type=String,Description="Ancestral Allele">
 ##INFO=<ID=DB,Number=0,Type=Flag,Description="dbSNP membership, build 129">
 ##INFO=<ID=H2,Number=0,Type=Flag,Description="HapMap2 membership">
+##INFO=<ID=ANNO,Number=.,Type=String,Description="Additional annotation">
 ##FILTER=<ID=q10,Description="Quality below 10">
 ##FILTER=<ID=s50,Description="Less than 50% of samples have data">
 ##FORMAT=<ID=GT,Number=1,Type=String,Description="Genotype">
 ##FORMAT=<ID=GQ,Number=1,Type=Integer,Description="Genotype Quality">
 ##FORMAT=<ID=DP,Number=1,Type=Integer,Description="Read Depth">
 ##FORMAT=<ID=HQ,Number=2,Type=Integer,Description="Haplotype Quality">
-##FORMAT=<ID=FT,Number=.,Type=String,Description="Haplotype Quality">
+##FORMAT=<ID=FT,Number=.,Type=String,Description="Call-wise filters">
 #CHROM\tPOS\tID\tREF\tALT\tQUAL\tFILTER\tINFO\tFORMAT\tNA00001\tNA00002\tNA00003
 """.lstrip()
 
@@ -51,5 +52,24 @@ def test_parse_minimal_record():
         "[Call('NA00001', OrderedDict([('GT', '0/1')])),"
         " Call('NA00002', OrderedDict([('GT', '0/2')])),"
         " Call('NA00003', OrderedDict([('GT', None)]))])")
+    RESULT = p.parse_next_record()
+    assert str(RESULT) == EXPECTED
+
+
+def test_parse_record_with_escaping():
+    # Setup parser with stock header and lines to parse
+    LINES = ('20\t100\t.\tC\tG\t.\t.\tANNO=Here%2Care%25some chars,'
+             '%2525\tGT:FT\t0/1:%2525,FOO\t0/0:.\t1/1:.\n')
+    p = vcf_parser(LINES)
+    hdr = p.parse_header()
+    # Perform the actual test
+    EXPECTED = (
+        "Record('20', 100, [], 'C', [Substitution(type_='SNV', value='G')], "
+        "None, [], OrderedDict([('ANNO', ['Here,are%some chars', '%25'])]), "
+        "['GT', 'FT'], "
+        "[Call('NA00001', OrderedDict([('GT', '0/1'),"
+        " ('FT', ['%25', 'FOO'])])),"
+        " Call('NA00002', OrderedDict([('GT', '0/0'), ('FT', [])])),"
+        " Call('NA00003', OrderedDict([('GT', '1/1'), ('FT', [])]))])")
     RESULT = p.parse_next_record()
     assert str(RESULT) == EXPECTED