+++ vcfpy/parser.py
@@ -11,6 +11,7 @@ from . import header
 from . import record
 from . import exceptions
 
+# TODO: add sphinx documentation for this
 try:
     from cyordereddict import OrderedDict
 except ImportError:
@@ -36,12 +37,20 @@ def _warn(msg):
     print('[vcfpy] WARNING: {}'.format(msg), file=sys.stderr)
 
 
-def split_quoted_string(s, delim=',', quote='"'):
-    """Split string ``s`` at delimiter, correctly interpreting quotes"""
+def split_quoted_string(s, delim=',', quote='"', brackets='[]'):
+    """Split string ``s`` at delimiter, correctly interpreting quotes
+
+    Further, interprets arrays wrapped in one level of ``[]``.  No recursive
+    brackets are interpreted (as this would make the grammar non-regular and
+    currently this complexity is not needed).  Currently, quoting inside of
+    braces is not supported either.  This is just to support the example
+    from VCF v4.3.
+    """
+    assert len(brackets) == 2
     # collect positions
     begins, ends = [0], []
     # run state automaton
-    NORMAL, QUOTED, ESCAPED, DELIM = 0, 1, 2, 3
+    NORMAL, QUOTED, ESCAPED, ARRAY, DELIM = 0, 1, 2, 3, 4
     state = NORMAL
     for pos, c in enumerate(s):
         if state == NORMAL:
@@ -50,12 +59,19 @@ def split_quoted_string(s, delim=',', quote='"'):
                 state = DELIM
             elif c == quote:
                 state = QUOTED
+            elif c == brackets[0]:
+                state = ARRAY
             else:
                 pass  # noop
         elif state == QUOTED:
             if c == '\\':
                 state = ESCAPED
             elif c == quote:
+                state = NORMAL
+            else:
+                pass  # noop
+        elif state == ARRAY:
+            if c == brackets[1]:
                 state = NORMAL
             else:
                 pass  # noop
@@ -95,6 +111,8 @@ def parse_mapping(value):
             key = key.strip()  # XXX lenient parsing
             if value.startswith('"') and value.endswith('"'):
                 value = ast.literal_eval(value)
+            elif value.startswith('[') and value.endswith(']'):
+                value = [v.strip() for v in value[1:-1].split(',')]
         else:
             key, value = pair, True
         key_values.append((key, value))
@@ -137,17 +155,15 @@ class MappingHeaderLineParser(HeaderLineParserBase):
 
 # Parsers to use for each VCF header type (given left of '=')
 HEADER_PARSERS = {
+    'ALT': MappingHeaderLineParser(header.AltAlleleHeaderLine),
+    'contig': MappingHeaderLineParser(header.ContigHeaderLine),
     'FILTER': MappingHeaderLineParser(header.FilterHeaderLine),
     'FORMAT': MappingHeaderLineParser(header.FormatHeaderLine),
     'INFO': MappingHeaderLineParser(header.InfoHeaderLine),
-    'contig': MappingHeaderLineParser(header.ContigHeaderLine),
+    'META': MappingHeaderLineParser(header.MetaHeaderLine),
+    'PEDIGREE': MappingHeaderLineParser(header.PedigreeHeaderLine),
+    'SAMPLE': MappingHeaderLineParser(header.SampleHeaderLine),
     '__default__': StupidHeaderLineParser(),  # fallback
-    # 'ALT': None,
-    # 'assembly': None,
-    # 'META': None,
-    # 'SAMPLE': None,
-    # 'PEDIGREE': None,
-    # 'pedigreeDB': None,
 }
 
 
@@ -165,6 +181,11 @@ def convert_field_value(key, type_, value):
     """Convert atomic field value according to the type"""
     if value == '.':
         return None
+    elif type_ in ('Character', 'String'):
+        if '%' in value:
+            for k, v in record.UNESCAPE_MAPPING:
+                value = value.replace(k, v)
+        return value
     else:
         return _CONVERTERS[type_](value)
 
@@ -274,6 +295,8 @@ class RecordParser:
             self.expected_fields = 8
         # Cache of FieldInfo objects by FORMAT string
         self._format_cache = {}
+        # Cache of FILTER entries, also applied to FORMAT/FT
+        self._filter_ids = set(self.header.filter_ids())
 
     def parse_line(self, line_str):
         """Parse line from file (including trailing line break) and return
@@ -312,6 +335,7 @@ class RecordParser:
             filt = []
         else:
             filt = arr[6].split(';')
+        self._check_filters(filt, 'FILTER')
         # INFO
         info = self._parse_info(arr[7])
         if not self.samples.names:
@@ -329,8 +353,25 @@ class RecordParser:
                      zip(self.samples.names,
                          self._parse_calls_data(
                              format, self._format_cache[format_str], arr[9:]))]
+            for call in calls:
+                self._check_filters(
+                    call.data.get('FT'), 'FORMAT/FT', call.sample)
         return record.Record(
             chrom, pos, ids, ref, alts, qual, filt, info, format, calls)
+
+    def _check_filters(self, filt, source, sample=None):
+        if not filt:
+            return
+        for f in filt:
+            if f not in self._filter_ids:
+                if source == 'FILTER':
+                    _warn(('Filter not found in header: {}; found in '
+                           'FILTER column').format(f))
+                else:
+                    assert source == 'FORMAT/FT' and sample
+                    _warn(('Filter not found in header: {}; found in '
+                           'FORMAT/FT column of sample {}').format(
+                               f, sample))
 
     def _split_line(self, line_str):
         """Split line and check number of columns"""